chore: update client examples
