+++ packages/client/examples/node.js/fixtures/a.txt
@@ -1,0 +1,1 @@
+hello world
